+++ tests/Classes/Request.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace PlugRoute\Test\Classes;
+
+class Request extends \PlugRoute\Http\Request
+{
+	public function __construct()
+	{
+	}
+}