+++ doc/changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### v3.2
+* Improvements in the route group
+* Added redirect route
+* Improvement when obtaining data of the formdata type
+* Added parameter on construct
+
 ### v3.1
 * Added namespace method
 * Improvements in the route flow